pyproject.toml
[project]
name = "clickbait-detector"
version = "0.1.0"
description = "Clickbait Detection using a machine learning model"
dependencies = [
    "pandas==2.2.2",
    "transformers==4.40.1",
    "mlflow==2.13.2"
]

[project.optional-dependencies]
dev = [
    "torchinfo",
    "flake8"
]

[build-system]
build-backend = "flit_core.buildapi"
requires = ["flit_core >=3.2,<4"]